requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.15
django-filter==24.3
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.7
packaging==24.1
pillow==10.4.0
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
vk-api==11.9.9
vkapi==1.1
